// src/dora/companies/constants/companies.enum.ts
export enum ECompanies {
  AFAS = 'AFAS',
  ALPHABET = 'ALPHABET',
  BITVAVO = 'BITVAVO',
  BITWARDEN = 'BITWARDEN',
  BLOCKCHAINDOTCOM = 'BLOCKCHAINDOTCOM',
  BLOCKRISE = 'BLOCKRISE',
  BLOCKRISE_GROUP = 'BLOCKRISE_GROUP',
  STICHTING_BLOCKRISE = 'STICHTING_BLOCKRISE',
  BUNQ = 'BUNQ',
  CHAINSTACK = 'CHAINSTACK',
  CLOUDFLARE = 'CLOUDFLARE',
  COINGECKO = 'COINGECKO',
  COLT = 'COLT',
  COMPLYADVANTAGE = 'COMPLYADVANTAGE',
  ELLIPTIC = 'ELLIPTIC',
  GITHUB = 'GITHUB',
  GLASSNODE = 'GLASSNODE', // Only used by Corporate for chart analysis
  GRAFANA = 'GRAFANA',
  GOOGLE = 'GOOGLE', // Google Cloud ?
  GOOGLE_IRELAND = 'GOOGLE_IRELAND',
  HEXNODE = 'HEXNODE',
  IDIN = 'IDIN', // Bluem
  ING = 'ING',
  KRAKEN = 'KRAKEN',
  KWINZO = 'KWINZO',
  LINEAR = 'LINEAR',
  MEMPOOL = 'MEMPOOL',
  MICROSOFT = 'MICROSOFT',
  NEXTNOVATE = 'NEXTNOVATE', // Only as "Betaalconstructie" for payment to Google Cloud. If Nextnovate falls out then we can still pay directly to Google.
  NOTION = 'NOTION',
  POSTCODEAPI = 'POSTCODEAPI',
  PAY_DOT = 'PAY_DOT',
  QUICKNODE = 'QUICKNODE',
  SANITY = 'SANITY',
  SCRIVE = 'SCRIVE',
  SECUROSYS = 'SECUROSYS',
  SENDGRID = 'SENDGRID',
  SENTRY = 'SENTRY',
  SHIFTBASE = 'SHIFTBASE',
  SLACK = 'SLACK',
  SONARCLOUD = 'SONARCLOUD',
  SUREPAY = 'SUREPAY', // Bluem
  TWILIO = 'TWILIO',
  UBIQUITI = 'UBIQUITI', // Unifi
  UNISCAPE = 'UNISCAPE',
  VERCEL = 'VERCEL',
  VERIFF = 'VERIFF',
  ZENDESK = 'ZENDESK',
}

// Vercel
// Google Cloud
// Cloudflare
// Securosys
// Bitwarden
// Postcode API
// Scrive
// Bluem (IDIN + SurePay)
// Veriff
// ING
// Pay.
//   Twilio
// Sanity
// Zendesk
// Sendgrid
// Shiftbase
// Notion
// Kwinzo
// Microsoft
// GRC Tool
// ComplyAdvantage
// Grafana
// Glassnode
// Mempool
// Blockchain.com
// Bitvavo
// Kraken
// Ubiquity
// Hexnode
// AFAS
// Bunq
// Elliptic
// Chainstack





// COINGECKO
// COLT
// GITHUB
// LINEAR
// SENTRY
// SLACK
// SONARCLOUD




// BLOCKCHAINDOTCOM = 'BLOCKCHAINDOTCOM'
// GLASSNODE = 'GLASSNODE'
// GRAFANA = 'GRAFANA'
// GRCTOOL = 'GRCTOOL'
// HEXNODE = 'HEXNODE'
// ING = 'ING'
// KWINZO = 'KWINZO'
// MEMPOOL = 'MEMPOOL'
// SHIFTBASE  = 'SHIFT'
// ZENDESK = 'ZENDESK'